fixup

+++ src/logging.h
@@ -36,41 +36,41 @@ struct LogCategory {
 };
 
 namespace BCLog {
-    enum LogFlags : uint32_t {
+    enum LogFlags : uint64_t {
         NONE        = 0,
-        NET         = (1 <<  0),
-        TOR         = (1 <<  1),
-        MEMPOOL     = (1 <<  2),
-        HTTP        = (1 <<  3),
-        BENCH       = (1 <<  4),
-        ZMQ         = (1 <<  5),
-        WALLETDB    = (1 <<  6),
-        RPC         = (1 <<  7),
-        ESTIMATEFEE = (1 <<  8),
-        ADDRMAN     = (1 <<  9),
-        SELECTCOINS = (1 << 10),
-        REINDEX     = (1 << 11),
-        BLOCK       = (1 << 12),
-        RAND        = (1 << 13),
-        PRUNE       = (1 << 14),
-        PROXY       = (1 << 15),
-        MEMPOOLREJ  = (1 << 16),
-        LIBEVENT    = (1 << 17),
-        COINDB      = (1 << 18),
-        QT          = (1 << 19),
-        LEVELDB     = (1 << 20),
-        VALIDATION  = (1 << 21),
-        I2P         = (1 << 22),
-        IPC         = (1 << 23),
-        LOCK        = (1 << 24),
-        UTIL        = (1 << 25),
-        BLOCKSTORE  = (1 << 26),
-        CONN        = (1 << 27),
-        BLOCKSEND   = (1 << 28),
-        BLOCKBLOCK  = (1 << 29),
-        ANYONECANSPEND = (1 << 30),
-        BANMAN         = (1 << 31),
-        ALL         = ~(uint32_t)0,
+        NET         = (1ULL <<  0),
+        TOR         = (1ULL <<  1),
+        MEMPOOL     = (1ULL <<  2),
+        HTTP        = (1ULL <<  3),
+        BENCH       = (1ULL <<  4),
+        ZMQ         = (1ULL <<  5),
+        WALLETDB    = (1ULL <<  6),
+        RPC         = (1ULL <<  7),
+        ESTIMATEFEE = (1ULL <<  8),
+        ADDRMAN     = (1ULL <<  9),
+        SELECTCOINS = (1ULL << 10),
+        REINDEX     = (1ULL << 11),
+        BLOCK       = (1ULL << 12),
+        RAND        = (1ULL << 13),
+        PRUNE       = (1ULL << 14),
+        PROXY       = (1ULL << 15),
+        MEMPOOLREJ  = (1ULL << 16),
+        LIBEVENT    = (1ULL << 17),
+        COINDB      = (1ULL << 18),
+        QT          = (1ULL << 19),
+        LEVELDB     = (1ULL << 20),
+        VALIDATION  = (1ULL << 21),
+        I2P         = (1ULL << 22),
+        IPC         = (1ULL << 23),
+        LOCK        = (1ULL << 24),
+        UTIL        = (1ULL << 25),
+        BLOCKSTORE  = (1ULL << 26),
+        CONN        = (1ULL << 27),
+        BLOCKSEND   = (1ULL << 28),
+        BLOCKBLOCK  = (1ULL << 29),
+        ANYONECANSPEND = (1ULL << 30),
+        BANMAN         = (1ULL << 31),
+        ALL         = ~(uint64_t)0,
     };
 
     class Logger
@@ -90,7 +90,7 @@ namespace BCLog {
         std::atomic_bool m_started_new_line{true};
 
         /** Log categories bitfield. */
-        std::atomic<uint32_t> m_categories{0};
+        std::atomic<uint64_t> m_categories{0};
 
         std::string LogTimestampStr(const std::string& str);
